refactor: change authOtions location
